Update projectroadmap.md
I didn't translate the requirement of downloading numpy and pandas as a serious requirement because of how it was phrased in the original requirement's documentation. It was clear after trying to run step 2 that both numpy and pandas were needed so those were added to the projectroadmap and Numpy has been downloaded. Now, to Pandas